Get blink_simple to load from dev container
blink_simple is a C program so there is definitely more work to go. Also I had to like, restart the container a couple of times to make sure it started in privellaged mode.

Also, I had to hold in the bootsel button while plugging in the pico. tbh I'm not sure if that is required to be able to flash code to the pico, but it took a while.

Also since picotool isn't on my path (or linux's equivalent to path), I had to use the picotool that lives at:

`sudo /home/vscode/.pico-sdk/picotool/2.0.0/picotool/picotool load /workspaces/myFirstBlinky/blink_simple/build/blink_simple.elf -fx `

To get it to work.

Still tho. Very noice!